+++ include/modules/docker/DockerStatsDialog.h
@@ -5,9 +5,16 @@
 #ifndef AWSMOCK_QT_UI_DOCKER_STATS_DIALOG_H
 #define AWSMOCK_QT_UI_DOCKER_STATS_DIALOG_H
 
+#include <QMenu>
 #include <QDialog>
+#include <QList>
 
 #include <modules/docker/DockerService.h>
+#include <utils/IconUtils.h>
+#include <utils/BaseDialog.h>
+#include <modules/application/ApplicationEditDialog.h>
+#include <modules/application/ApplicationUploadCodeDialog.h>
+#include <utils/PrefixFilterModel.h>
 
 QT_BEGIN_NAMESPACE
 
@@ -17,7 +24,7 @@ namespace Ui {
 
 QT_END_NAMESPACE
 
-class DockerStatsDialog final : public QDialog {
+class DockerStatsDialog final : public BaseDialog {
     Q_OBJECT
 
 public:
@@ -29,7 +36,13 @@ public:
 
     void HandleReject();
 
-    void LoadContent(const DockerStatsResponse &dockerStatsResponse);
+    double GetCpuPercent(const ContainerStat &containerStats);
+
+    void ShowContextMenu(const QPoint &pos);
+
+    void LoadContainerStatsContent(const DockerStatsResponse &dockerStatsResponse);
+
+    void LoadContent() override;
 
 private:
     /**
@@ -41,6 +54,56 @@ private:
      * @brief Container service
      */
     DockerService *_containerService;
+
+    /**
+     * @brief Application service
+     */
+    ApplicationService *_applicationService;
+
+    /**
+     * @brief Sort column index
+     */
+    int _sortColumn = 0;
+
+    /**
+     * @brief Sort order
+     */
+    Qt::SortOrder _sortOrder = Qt::AscendingOrder;
+
+    /**
+     * @brief List of container IDs
+     */
+    QList<QString> _containerIds;
+
+    /**
+     * @brief Prefix suche
+     */
+    QString _prefixValue = "";
+
+    /**
+     * @brief Previous CPU total value
+     */
+    std::map<QString, long> _oldCpuTotal{};
+
+    /**
+     * @brief Previous CPU system value
+     */
+    std::map<QString, long> _oldCpuSystem{};
+
+    /**
+     *  @brief Table data model
+     */
+    QStandardItemModel *_dataModel;
+
+    /**
+     * @brief Data proxy model
+     */
+    PrefixFilterProxyModel *_proxyModel;
+
+    /**
+     * @brief Status bar connection
+     */
+    QMetaObject::Connection _statusConnection;
 };